ddi-langgraph/requirements.txt
langgraph>=0.2.30
langgraph-cli>=0.1.38
langchain-core>=0.3.10
langchain-openai>=0.2.4
langgraph-sdk>=0.1.12
neo4j>=5.24.0
python-dotenv>=1.0.1
pandas>=2.2.2
